refactor(core): originate random number always from sec/rng
[no changelog]
