+++ core/SConscript.bootloader_emu
@@ -68,6 +68,7 @@ SOURCE_MOD_CRYPTO += [
     'vendor/trezor-crypto/ed25519-donna/modm-donna-32bit.c',
     'vendor/trezor-crypto/memzero.c',
     'vendor/trezor-crypto/rand.c',
+    'vendor/trezor-crypto/rand_insecure.c',
     'vendor/trezor-crypto/sha2.c',
 ]
