+++ core/SConscript.prodtest_emu
@@ -87,6 +87,7 @@ SOURCE_MOD_CRYPTO += [
     'vendor/trezor-crypto/noise.c',
     'vendor/trezor-crypto/nist256p1.c',
     'vendor/trezor-crypto/rand.c',
+    'vendor/trezor-crypto/rand_insecure.c',
     'vendor/trezor-crypto/ripemd160.c',
     'vendor/trezor-crypto/rfc6979.c',
     'vendor/trezor-crypto/secp256k1.c',
