+++ crypto/rand.c
@@ -23,47 +23,6 @@
 
 #include "rand.h"
 
-#ifdef USE_INSECURE_PRNG
-
-#pragma message( \
-    "NOT SUITABLE FOR PRODUCTION USE! Replace random32() function with your own secure code.")
-
-// The following code is not supposed to be used in a production environment.
-// It's included only to make the library testable.
-// The message above tries to prevent any accidental use outside of the test
-// environment.
-//
-// You are supposed to replace the random8() and random32() function with your
-// own secure code. There is also a possibility to replace the random_buffer()
-// function as it is defined as a weak symbol.
-
-static uint32_t seed = 0;
-
-void random_reseed(const uint32_t value) { seed = value; }
-
-uint32_t random32(void) {
-  // Linear congruential generator from Numerical Recipes
-  // https://en.wikipedia.org/wiki/Linear_congruential_generator
-  seed = 1664525 * seed + 1013904223;
-  return seed;
-}
-
-#endif /* USE_INSECURE_PRNG */
-
-//
-// The following code is platform independent
-//
-
-void __attribute__((weak)) random_buffer(uint8_t *buf, size_t len) {
-  uint32_t r = 0;
-  for (size_t i = 0; i < len; i++) {
-    if (i % 4 == 0) {
-      r = random32();
-    }
-    buf[i] = (r >> ((i % 4) * 8)) & 0xFF;
-  }
-}
-
 void random_xor(uint8_t *buf, size_t len) {
   uint8_t r[4] = {0};
   for (size_t i = 0; i < len; i++) {
