+++ storage/tests/c/Makefile
@@ -22,6 +22,7 @@ SRC += storage/storage_utils.c
 SRC += storage/norcow.c
 SRC += crypto/pbkdf2.c
 SRC += crypto/rand.c
+SRC += crypto/rand_insecure.c
 SRC += crypto/chacha20poly1305/rfc7539.c
 SRC += crypto/chacha20poly1305/chacha20poly1305.c
 SRC += crypto/chacha20poly1305/poly1305-donna.c
@@ -50,11 +51,23 @@ build_qw/crypto/chacha20poly1305/chacha_merged.o: $(BASE)crypto/chacha20poly1305
 	mkdir -p $(@D)
 	$(CC) $(CFLAGS) $(INC) -c $< -o $@
 
+# build object files from C sources with headers
 build/%.o: $(BASE)%.c $(BASE)%.h
 	mkdir -p $(@D)
 	$(CC) $(CFLAGS) -DFLASH_BIT_ACCESS -DFLASH_BLOCK_WORDS=1 $(INC) -c $< -o $@
 
+# build object files from C sources without headers
+build/%.o: $(BASE)%.c
+	mkdir -p $(@D)
+	$(CC) $(CFLAGS) -DFLASH_BIT_ACCESS -DFLASH_BLOCK_WORDS=1 $(INC) -c $< -o $@
+
+# build object files from C sources with headers
 build_qw/%.o: $(BASE)%.c $(BASE)%.h
+	mkdir -p $(@D)
+	$(CC) $(CFLAGS) -DFLASH_BLOCK_WORDS=4  $(INC) -c $< -o $@
+
+# build object files from C sources without headers
+build_qw/%.o: $(BASE)%.c
 	mkdir -p $(@D)
 	$(CC) $(CFLAGS) -DFLASH_BLOCK_WORDS=4  $(INC) -c $< -o $@
